third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/application-panel/preview-source-a11y-test-expected.txt

Tests accessibility in the preview view for source files using the axe-core linter.

Initial state:
Application
  Manifest (selected)
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
  Web SQL
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
  Application Cache
Background Services
  Background Fetch
  Background Sync
Frames
  top
visible view: unknown

Running: testSearchableView
Revealed preview-source-a11y-test.js:
Application
  Manifest
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
  Web SQL
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
  Application Cache
Background Services
  Background Fetch
  Background Sync
Frames
  top
    Scripts
      preview-source-a11y-test.js (selected)
    inspected-page.html
visible view: source
aXe violations: []
